cleanups

+++ userspace/libsinsp/scores.cpp
@@ -757,7 +757,7 @@ float sinsp_scores::get_process_capacity_score(float system_capacity_score, sins
 	//
 	// Take the system health score and normalize it using the local/remote ratios
 	//
-	if(mainthread_info->m_procinfo->m_proc_transaction_metrics.m_counter.m_time_ns_in != 0)
+	if(mainthread_info->m_procinfo->m_proc_transaction_metrics.m_counter.m_time_ns_in != (uint64_t)0)
 	{
 		local_remote_ratio = (float)mainthread_info->m_procinfo->m_proc_transaction_processing_delay_ns / 
 			(float)mainthread_info->m_procinfo->m_proc_transaction_metrics.m_counter.m_time_ns_in;
